refactor(agent): phase B — decompose honeypot/mod.rs
