+++ crates/agent/src/skills/builtin/honeypot/audit.rs
@@ -1,0 +1,2 @@
+//! Phase B decomposition module.
+//! Audit writer and forensics lifecycle helpers will live here.
