+++ crates/agent/src/skills/builtin/honeypot/banner.rs
@@ -1,0 +1,2 @@
+//! Phase B decomposition module.
+//! Banner and endpoint helpers will live here.
