+++ crates/agent/src/skills/builtin/honeypot/containment.rs
@@ -1,0 +1,2 @@
+//! Phase B decomposition module.
+//! Containment helpers (namespace/jail/process) will live here.
