+++ crates/agent/src/skills/builtin/honeypot/pcap_handoff.rs
@@ -1,0 +1,2 @@
+//! Phase B decomposition module.
+//! PCAP and external handoff helpers will live here.
